Add BasicAuth utility function
This commit adds a function called `BasicAuth` which accepts an API
Gateway proxy request, and attempts to extract username and password
provided via Basic HTTP Authentication in the "Authorization" header.

If the header does not exists, or does not contain valid Basic
authentication data, the function will return empty values.