[fix #3] overlapping routes may not be matched
When a router includes two routes that overlap, such as `GET /foo/:id`
and `POST /foo/bar`, requests for the second route may sometimes fail
with a `405 Method Not Allowed` response. This happens because the
router will sometimes match the first route instead of the second one,
due to the non-deterministic nature of the route map.

This commit fixes the issue by continuing the matching process if the
request method is not supported by the first matched route.