Update documentation
The README file is updated to remove the notice about the library being
in alpha status, which hasn't been true for a long time. Some more
information is also added.

The embedded documentation in decoder.go is also updated with an example
of a struct type that takes parameters from both the body and other
sources.